apply change them
