+++ src/lib/mui/palette.ts
@@ -1,10 +1,8 @@
-import { alpha } from '@mui/material';
+import { alpha, Palette } from '@mui/material';
+import GREEN_PALETTE from './paletteList/green';
+import PURPLE_PALETTE from './paletteList/purple';
 
 // ----------------------------------------------------------------------
-
-function createGradient(color1: string, color2: string) {
-  return `linear-gradient(to bottom, ${color1}, ${color2})`;
-}
 
 interface GradientsPaletteOptions {
   primary: string;
@@ -48,111 +46,41 @@ declare module '@mui/material' {
   }
 }
 
-// SETUP COLORS
-const PRIMARY = {
-  lighter: '#C8FACD',
-  light: '#5BE584',
-  main: '#00AB55',
-  dark: '#007B55',
-  darker: '#005249',
-};
-const SECONDARY = {
-  lighter: '#D6E4FF',
-  light: '#84A9FF',
-  main: '#3366FF',
-  dark: '#1939B7',
-  darker: '#091A7A',
-};
-const INFO = {
-  lighter: '#D0F2FF',
-  light: '#74CAFF',
-  main: '#1890FF',
-  dark: '#0C53B7',
-  darker: '#04297A',
-};
-const SUCCESS = {
-  lighter: '#E9FCD4',
-  light: '#AAF27F',
-  main: '#54D62C',
-  dark: '#229A16',
-  darker: '#08660D',
-};
-const WARNING = {
-  lighter: '#FFF7CD',
-  light: '#FFE16A',
-  main: '#FFC107',
-  dark: '#B78103',
-  darker: '#7A4F01',
-};
-const ERROR = {
-  lighter: '#FFE7D9',
-  light: '#FFA48D',
-  main: '#FF4842',
-  dark: '#B72136',
-  darker: '#7A0C2E',
-};
-const GREY = {
-  0: '#FFFFFF',
-  100: '#F9FAFB',
-  200: '#F4F6F8',
-  300: '#DFE3E8',
-  400: '#C4CDD5',
-  500: '#919EAB',
-  600: '#637381',
-  700: '#454F5B',
-  800: '#212B36',
-  900: '#161C24',
-  500_8: alpha('#919EAB', 0.08),
-  500_12: alpha('#919EAB', 0.12),
-  500_16: alpha('#919EAB', 0.16),
-  500_24: alpha('#919EAB', 0.24),
-  500_32: alpha('#919EAB', 0.32),
-  500_48: alpha('#919EAB', 0.48),
-  500_56: alpha('#919EAB', 0.56),
-  500_80: alpha('#919EAB', 0.8),
-};
-const GRADIENTS = {
-  primary: createGradient(PRIMARY.light, PRIMARY.main),
-  info: createGradient(INFO.light, INFO.main),
-  success: createGradient(SUCCESS.light, SUCCESS.main),
-  warning: createGradient(WARNING.light, WARNING.main),
-  error: createGradient(ERROR.light, ERROR.main),
-};
-const COMMON = {
-  common: { black: '#000', white: '#fff' },
-  primary: { ...PRIMARY, contrastText: '#fff' },
-  secondary: { ...SECONDARY, contrastText: '#fff' },
-  info: { ...INFO, contrastText: '#fff' },
-  success: { ...SUCCESS, contrastText: GREY[800] },
-  warning: { ...WARNING, contrastText: GREY[800] },
-  error: { ...ERROR, contrastText: '#fff' },
-  grey: GREY,
-  gradients: GRADIENTS,
-  divider: GREY[500_24],
-  action: {
-    hover: GREY[500_8],
-    selected: GREY[500_16],
-    disabled: GREY[500_80],
-    disabledBackground: GREY[500_24],
-    focus: GREY[500_24],
-    hoverOpacity: 0.08,
-    disabledOpacity: 0.48,
-  },
+// const palette = {
+//   light: {
+//     ...COMMON,
+//     text: { primary: GREY[800], secondary: GREY[600], disabled: GREY[500] },
+//     background: { paper: '#fff', default: '#fff', neutral: GREY[200] },
+//     action: { active: GREY[600], ...COMMON.action },
+//   },
+//   dark: {
+//     ...COMMON,
+//     text: { primary: '#fff', secondary: GREY[500], disabled: GREY[600] },
+//     background: { paper: GREY[800], default: GREY[900], neutral: GREY[500_16] },
+//     action: { active: GREY[500], ...COMMON.action },
+//   },
+// };
+
+const PALETTE_LIST: { [key: string]: any } = {
+  green: GREEN_PALETTE,
+  purple: PURPLE_PALETTE,
 };
 
-const palette = {
-  light: {
-    ...COMMON,
-    text: { primary: GREY[800], secondary: GREY[600], disabled: GREY[500] },
-    background: { paper: '#fff', default: '#fff', neutral: GREY[200] },
-    action: { active: GREY[600], ...COMMON.action },
-  },
-  dark: {
-    ...COMMON,
-    text: { primary: '#fff', secondary: GREY[500], disabled: GREY[600] },
-    background: { paper: GREY[800], default: GREY[900], neutral: GREY[500_16] },
-    action: { active: GREY[500], ...COMMON.action },
-  },
-};
+export const getPalette = (paletteName: string) => {
+  const customPalette = PALETTE_LIST[paletteName];
 
-export default palette;
+  return {
+    light: {
+      ...customPalette.COMMON,
+      text: { primary: customPalette.GREY[800], secondary: customPalette.GREY[600], disabled: customPalette.GREY[500] },
+      background: { paper: '#fff', default: '#fff', neutral: customPalette.GREY[200] },
+      action: { active: customPalette.GREY[600], ...customPalette.COMMON.action },
+    },
+    dark: {
+      ...customPalette.COMMON,
+      text: { primary: '#fff', secondary: customPalette.GREY[500], disabled: customPalette.GREY[600] },
+      background: { paper: customPalette.GREY[800], default: customPalette.GREY[900], neutral: customPalette.GREY[500_16] },
+      action: { active: customPalette.GREY[500], ...customPalette.COMMON.action },
+    },
+  };
+}
