+++ src/lib/mui/paletteList/purple.ts
@@ -1,0 +1,112 @@
+import { alpha } from '@mui/material';
+
+function createGradient(color1: string, color2: string) {
+  return `linear-gradient(to bottom, ${color1}, ${color2})`;
+}
+
+// SETUP COLORS
+const PRIMARY = {
+  lighter: '#EBD6FD',
+  light: '#B985F4',
+  main: '#7635dc',
+  dark: '#431A9E',
+  darker: '#200A69',
+};
+const SECONDARY = {
+  lighter: '#FDF2D6',
+  light: '#F5CB83',
+  main: '#e09032',
+  dark: '#A15619',
+  darker: '#6B2C09',
+};
+const INFO = {
+  lighter: '#D0F2FF',
+  light: '#74CAFF',
+  main: '#1890FF',
+  dark: '#0C53B7',
+  darker: '#04297A',
+};
+const SUCCESS = {
+  lighter: '#E9FCD4',
+  light: '#AAF27F',
+  main: '#54D62C',
+  dark: '#229A16',
+  darker: '#08660D',
+};
+const WARNING = {
+  lighter: '#FFF7CD',
+  light: '#FFE16A',
+  main: '#FFC107',
+  dark: '#B78103',
+  darker: '#7A4F01',
+};
+const ERROR = {
+  lighter: '#FFE7D9',
+  light: '#FFA48D',
+  main: '#FF4842',
+  dark: '#B72136',
+  darker: '#7A0C2E',
+};
+const GREY = {
+  0: '#FFFFFF',
+  100: '#F9FAFB',
+  200: '#F4F6F8',
+  300: '#DFE3E8',
+  400: '#C4CDD5',
+  500: '#919EAB',
+  600: '#637381',
+  700: '#454F5B',
+  800: '#212B36',
+  900: '#161C24',
+  500_8: alpha('#919EAB', 0.08),
+  500_12: alpha('#919EAB', 0.12),
+  500_16: alpha('#919EAB', 0.16),
+  500_24: alpha('#919EAB', 0.24),
+  500_32: alpha('#919EAB', 0.32),
+  500_48: alpha('#919EAB', 0.48),
+  500_56: alpha('#919EAB', 0.56),
+  500_80: alpha('#919EAB', 0.8),
+};
+const GRADIENTS = {
+  primary: createGradient(PRIMARY.light, PRIMARY.main),
+  info: createGradient(INFO.light, INFO.main),
+  success: createGradient(SUCCESS.light, SUCCESS.main),
+  warning: createGradient(WARNING.light, WARNING.main),
+  error: createGradient(ERROR.light, ERROR.main),
+};
+const COMMON = {
+  common: { black: '#000', white: '#fff' },
+  primary: { ...PRIMARY, contrastText: '#fff' },
+  secondary: { ...SECONDARY, contrastText: '#fff' },
+  info: { ...INFO, contrastText: '#fff' },
+  success: { ...SUCCESS, contrastText: GREY[800] },
+  warning: { ...WARNING, contrastText: GREY[800] },
+  error: { ...ERROR, contrastText: '#fff' },
+  grey: GREY,
+  gradients: GRADIENTS,
+  divider: GREY[500_24],
+  action: {
+    hover: GREY[500_8],
+    selected: GREY[500_16],
+    disabled: GREY[500_80],
+    disabledBackground: GREY[500_24],
+    focus: GREY[500_24],
+    hoverOpacity: 0.08,
+    disabledOpacity: 0.48,
+  },
+};
+
+const PURPLE_PALETTE = {
+  // SETUP COLORS
+  PRIMARY: PRIMARY,
+  SECONDARY: SECONDARY,
+  INFO: INFO,
+  SUCCESS: SUCCESS,
+  WARNING: WARNING,
+  ERROR: ERROR,
+  GREY: GREY,
+  GRADIENTS: GRADIENTS,
+  COMMON: COMMON,
+}
+
+export default PURPLE_PALETTE;
