+++ src/redux/slices/currentTenantSlice.ts
@@ -5,16 +5,19 @@ import { RootState } from '../store';
 export type Tenant = {
   subdomain: string;
   tenantId: string;
+  theme: string;
 };
 
 export interface CurrentTenantState {
   subdomain: string;
   tenantId: string;
+  theme: string;
 }
 
 const initialState: CurrentTenantState = {
   subdomain: '',
   tenantId: '',
+  theme: 'green',
 };
 
 const currentTenantSlice = createSlice({
@@ -32,5 +35,8 @@ export const { setTenant } = currentTenantSlice.actions;
 
 export const currentTenantSubdomainSelector = (state: RootState) =>
   state.currentTenant.subdomain;
+
+export const currentTenantThemeSelector = (state: RootState) =>
+  state.currentTenant.theme;
 
 export default currentTenantSlice.reducer;
